moved loader/ into models/
