+++ fastvideo/v1/models/encoders/llama.py
@@ -41,7 +41,7 @@ from fastvideo.v1.layers.linear import (MergedColumnParallelLinear,
 
 from fastvideo.v1.layers.rotary_embedding import get_rope
 from fastvideo.v1.layers.vocab_parallel_embedding import VocabParallelEmbedding
-from fastvideo.v1.loader.weight_utils import (
+from fastvideo.v1.models.loader.weight_utils import (
     default_weight_loader, maybe_remap_kv_scale_name)
 
 from .utils import (extract_layer_index)
