+++ fastvideo/v1/tests/test_llama_encoder.py
@@ -84,7 +84,7 @@ def test_llama_encoder():
     )
 
     from fastvideo.v1.models.component_loader import TextEncoderLoader
-    from fastvideo.v1.loader.loader import DefaultModelLoader
+    from fastvideo.v1.models.loader.loader import DefaultModelLoader
     loader = DefaultModelLoader()
     args.device_str = "cuda:0"
     model2 = loader.load_model(model_path, hf_config, args)
